refactor: move REST attributes to controller
This continues the cleanup refactor with the Budgets.
