+++ pkg/models/budget.go
@@ -17,17 +17,6 @@ import (
 type Budget struct {
 	DefaultModel
 	BudgetCreate
-	Balance decimal.Decimal `json:"balance" gorm:"-" example:"3423.42"` // DEPRECATED. Will be removed in API v2, see https://github.com/envelope-zero/backend/issues/526.
-	Links   struct {
-		Self             string `json:"self" example:"https://example.com/api/v1/budgets/550dc009-cea6-4c12-b2a5-03446eb7b7cf"`                                 // The budget itself
-		Accounts         string `json:"accounts" example:"https://example.com/api/v1/accounts?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf"`                     // Accounts for this budget
-		Categories       string `json:"categories" example:"https://example.com/api/v1/categories?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf"`                 // Categories for this budget
-		Envelopes        string `json:"envelopes" example:"https://example.com/api/v1/envelopes?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf"`                   // Envelopes for this budget
-		Transactions     string `json:"transactions" example:"https://example.com/api/v1/transactions?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf"`             // Transactions for this budget
-		Month            string `json:"month" example:"https://example.com/api/v1/budgets/550dc009-cea6-4c12-b2a5-03446eb7b7cf/YYYY-MM"`                        // This uses 'YYYY-MM' for clients to replace with the actual year and month.
-		GroupedMonth     string `json:"groupedMonth" example:"https://example.com/api/v1/months?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf&month=YYYY-MM"`     // This uses 'YYYY-MM' for clients to replace with the actual year and month.
-		MonthAllocations string `json:"monthAllocations" example:"https://example.com/api/v1/months?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf&month=YYYY-MM"` // This uses 'YYYY-MM' for clients to replace with the actual year and month.
-	} `json:"links" gorm:"-"`
 }
 
 func (b Budget) Self() string {
@@ -40,25 +29,8 @@ type BudgetCreate struct {
 	Currency string `json:"currency" example:"€" default:""`                // The currency for the budget
 }
 
-type BudgetMonth struct {
-	ID        uuid.UUID       `json:"id" example:"1e777d24-3f5b-4c43-8000-04f65f895578"` // The ID of the Budget
-	Name      string          `json:"name" example:"Groceries"`                          // The name of the Budget
-	Month     types.Month     `json:"month" example:"2006-05-01T00:00:00.000000Z"`       // Month these calculations are made for
-	Budgeted  decimal.Decimal `json:"budgeted" example:"2100"`                           // Amount of money that has been allocated to envelopes
-	Income    decimal.Decimal `json:"income" example:"2317.34"`                          // Income. This is all money that is sent from off-budget to on-budget accounts without an envelope set.
-	Available decimal.Decimal `json:"available" example:"217.34"`                        // The amount of money still available to budget.
-	Envelopes []EnvelopeMonth `json:"envelopes"`                                         // The envelopes this budget has, with detailed calculations
-}
-
-func (b BudgetMonth) Self() string {
-	return "BudgetMonth"
-}
-
-func (b *Budget) AfterFind(tx *gorm.DB) (err error) {
-	b.links(tx)
-
-	b.Balance = decimal.Zero
-
+// Balance calculates the balance for a budget.
+func (b Budget) Balance(tx *gorm.DB) (balance decimal.Decimal, err error) {
 	// Get all OnBudget accounts for the budget
 	var accounts []Account
 	_ = tx.Where(&Account{
@@ -70,33 +42,15 @@ func (b *Budget) AfterFind(tx *gorm.DB) (err error) {
 
 	// Add all their balances to the budget's balance
 	for _, account := range accounts {
-		balance, _, err := account.GetBalanceMonth(tx, types.Month{})
+		aBalance, _, err := account.GetBalanceMonth(tx, types.Month{})
 		if err != nil {
-			return err
+			return decimal.Zero, err
 		}
 
-		b.Balance = b.Balance.Add(balance)
+		balance = balance.Add(aBalance)
 	}
 
-	return
-}
-
-// AfterSave does the same as AfterFind, so we just call it.
-func (b *Budget) AfterSave(tx *gorm.DB) (err error) {
-	return b.AfterFind(tx)
-}
-
-func (b *Budget) links(tx *gorm.DB) {
-	url := tx.Statement.Context.Value(database.ContextURL)
-
-	b.Links.Self = fmt.Sprintf("%s/v1/budgets/%s", url, b.ID)
-	b.Links.Month = b.Links.Self + "/YYYY-MM"
-	b.Links.Accounts = fmt.Sprintf("%s/v1/accounts?budget=%s", url, b.ID)
-	b.Links.Categories = fmt.Sprintf("%s/v1/categories?budget=%s", url, b.ID)
-	b.Links.Envelopes = fmt.Sprintf("%s/v1/envelopes?budget=%s", url, b.ID)
-	b.Links.Transactions = fmt.Sprintf("%s/v1/transactions?budget=%s", url, b.ID)
-	b.Links.GroupedMonth = fmt.Sprintf("%s/v1/months?budget=%s&month=YYYY-MM", url, b.ID)
-	b.Links.MonthAllocations = fmt.Sprintf("%s/v1/months?budget=%s&month=YYYY-MM", url, b.ID)
+	return balance, nil
 }
 
 // Income returns the income for a budget in a given month.
@@ -149,27 +103,6 @@ func (b Budget) Allocated(db *gorm.DB, month types.Month) (allocated decimal.Dec
 	}
 
 	return
-}
-
-type CategoryEnvelopes struct {
-	Category
-	Envelopes  []EnvelopeMonth `json:"envelopes"`                // Slice of all envelopes
-	Balance    decimal.Decimal `json:"balance" example:"-10.13"` // Sum of the balances of the envelopes
-	Allocation decimal.Decimal `json:"allocation" example:"90"`  // Sum of allocations for the envelopes
-	Spent      decimal.Decimal `json:"spent" example:"100.13"`   // Sum spent for all envelopes
-}
-
-type Month struct {
-	ID         uuid.UUID           `json:"id" example:"1e777d24-3f5b-4c43-8000-04f65f895578"` // The ID of the Budget
-	Name       string              `json:"name" example:"Zero budget"`                        // The name of the Budget
-	Month      types.Month         `json:"month" example:"2006-05-01T00:00:00.000000Z"`       // The month
-	Budgeted   decimal.Decimal     `json:"budgeted" example:"2100"`                           // The sum of all allocations for the month. **Deprecated, please use the `allocation` field**
-	Income     decimal.Decimal     `json:"income" example:"2317.34"`                          // The total income for the month (sum of all incoming transactions without an Envelope)
-	Available  decimal.Decimal     `json:"available" example:"217.34"`                        // The amount available to budget
-	Balance    decimal.Decimal     `json:"balance" example:"5231.37"`                         // The sum of all envelope balances
-	Spent      decimal.Decimal     `json:"spent" example:"133.70"`                            // The amount of money spent in this month
-	Allocation decimal.Decimal     `json:"allocation" example:"1200.50"`                      // The sum of all allocations for this month
-	Categories []CategoryEnvelopes `json:"categories"`                                        // A list of envelope month calculations grouped by category
 }
 
 // Month calculates the month overview for this month.
